Changes to REST API Object Read
The newer versions of the KMIP Source Server have JSON extensions to the objects that do not exist in the older versions.  As such, object information collection had to be modified to use the older JSON key descriptors.  Furthermore, the returned values had to be parsed more surgically (using .split() .join(), title(), etc.).